feat: Add masks package
1. New package for attention masks
2. global_attention_mask.py
